@@ -1,47 +1,50 @@
 {
   "type": "project",
   "license": "proprietary",
+  "minimum-stability": "stable",
   "require": {
-    "php": "^7.4",
+    "php": "^8.0",
     "ext-ctype": "*",
     "ext-fileinfo": "*",
     "ext-iconv": "*",
     "composer/package-versions-deprecated": "^1.8",
     "doctrine/annotations": "^1.0",
-    "doctrine/doctrine-bundle": "^2.1",
-    "doctrine/doctrine-migrations-bundle": "^3.0",
-    "doctrine/orm": "^2.7",
+    "doctrine/doctrine-bundle": "2.5.*",
+    "doctrine/doctrine-migrations-bundle": "^3.2",
+    "doctrine/orm": "^2.9",
     "erusev/parsedown": "^1.7",
     "geertw/ip-anonymizer": "^1.1",
     "gumlet/php-image-resize": "^2.0",
     "phpdocumentor/reflection-docblock": "^5.1",
-    "sensio/framework-extra-bundle": "^5.1",
+    "sensio/framework-extra-bundle": "^6.2",
     "squizlabs/php_codesniffer": "^3.5",
-    "symfony/asset": "5.2.*",
-    "symfony/console": "5.2.*",
-    "symfony/dotenv": "5.2.*",
-    "symfony/expression-language": "5.2.*",
+    "symfony/asset": "5.4.*",
+    "symfony/console": "5.4.*",
+    "symfony/dotenv": "5.4.*",
+    "symfony/expression-language": "5.4.*",
     "symfony/flex": "^1.1",
-    "symfony/form": "5.2.*",
-    "symfony/framework-bundle": "5.2.*",
-    "symfony/http-client": "5.2.*",
-    "symfony/intl": "5.2.*",
-    "symfony/mailer": "5.2.*",
-    "symfony/mime": "5.2.*",
+    "symfony/form": "5.4.*",
+    "symfony/framework-bundle": "5.4.*",
+    "symfony/http-client": "5.4.*",
+    "symfony/intl": "5.4.*",
+    "symfony/mailer": "5.4.*",
+    "symfony/mime": "5.4.*",
     "symfony/monolog-bundle": "^3.1",
-    "symfony/notifier": "5.2.*",
-    "symfony/process": "5.2.*",
-    "symfony/property-access": "5.2.*",
-    "symfony/property-info": "5.2.*",
-    "symfony/security-bundle": "5.2.*",
-    "symfony/serializer": "5.2.*",
-    "symfony/string": "5.2.*",
-    "symfony/translation": "5.2.*",
-    "symfony/twig-bundle": "5.2.*",
-    "symfony/validator": "5.2.*",
-    "symfony/web-link": "5.2.*",
+    "symfony/notifier": "5.4.*",
+    "symfony/process": "5.4.*",
+    "symfony/property-access": "5.4.*",
+    "symfony/property-info": "5.4.*",
+    "symfony/security-bundle": "5.4.*",
+    "symfony/security-csrf": "5.4.*",
+    "symfony/serializer": "5.4.*",
+    "symfony/string": "5.4.*",
+    "symfony/translation": "5.4.*",
+    "symfony/twig-bundle": "5.4.*",
+    "symfony/uid": "5.4.*",
+    "symfony/validator": "5.4.*",
+    "symfony/web-link": "5.4.*",
     "symfony/webpack-encore-bundle": "^1.7",
-    "symfony/yaml": "5.2.*",
+    "symfony/yaml": "5.4.*",
     "twig/extra-bundle": "^2.12|^3.0",
     "twig/intl-extra": "^3.3",
     "twig/markdown-extra": "^3.2",
@@ -49,32 +52,38 @@
   },
   "require-dev": {
     "doctrine/doctrine-fixtures-bundle": "^3.3",
-    "liip/test-fixtures-bundle": "^1.0.0",
+    "friendsofphp/php-cs-fixer": "^2.15",
+    "liip/test-fixtures-bundle": "^2.3.0",
     "nelmio/alice": "^3.7",
-    "symfony/browser-kit": "^5.1",
-    "symfony/css-selector": "^5.1",
-    "symfony/debug-bundle": "^5.1",
+    "phpstan/phpstan": "^1.1",
+    "phpstan/phpstan-doctrine": "^1.0",
+    "phpstan/phpstan-mockery": "^1.0",
+    "phpstan/phpstan-phpunit": "^1.0",
+    "phpstan/phpstan-symfony": "^1.0",
+    "phpunit/phpunit": "^9.5",
+    "rector/rector": "^0.12.3",
+    "symfony/browser-kit": "^5.4",
+    "symfony/css-selector": "^5.4",
+    "symfony/debug-bundle": "^5.4",
     "symfony/maker-bundle": "^1.0",
     "symfony/monolog-bundle": "^3.0",
-    "symfony/phpunit-bridge": "^5.1",
-    "symfony/stopwatch": "^5.1",
-    "symfony/twig-bundle": "^5.1",
-    "symfony/var-dumper": "^5.1",
-    "symfony/web-profiler-bundle": "^5.1",
-    "theofidry/alice-data-fixtures": "^1.2",
-    "phpstan/phpstan": "^0.12.66",
-    "phpstan/phpstan-doctrine": "^0.12.27",
-    "phpstan/phpstan-mockery": "^0.12.11",
-    "phpstan/phpstan-phpunit": "^0.12.17",
-    "phpstan/phpstan-symfony": "^0.12.14",
-    "friendsofphp/php-cs-fixer": "^2.15"
+    "symfony/phpunit-bridge": "^5.4",
+    "symfony/stopwatch": "^5.4",
+    "symfony/twig-bundle": "^5.4",
+    "symfony/var-dumper": "^5.4",
+    "symfony/web-profiler-bundle": "^5.4",
+    "theofidry/alice-data-fixtures": "^1.4"
   },
   "config": {
     "optimize-autoloader": true,
     "preferred-install": {
       "*": "dist"
     },
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "composer/package-versions-deprecated": true,
+      "symfony/flex": true
+    }
   },
   "autoload": {
     "psr-4": {
