@@ -12,9 +12,9 @@ use App\Repository\Main\CV\RealisationRepository;
 use App\Utils\Assets\AssetsResponse;
 use App\Utils\StringHelper;
 use Doctrine\ORM\EntityManagerInterface;
+use Psr\Cache\CacheItemPoolInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Cache\Adapter\AdapterInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
@@ -27,19 +27,17 @@ class CuriculumVitaeController extends AbstractController
     public const CACHE_KEY_REALISATION_VIEW_GALLERY = 'cv_making_view_galley_';
     public const TECHNOLOGIES_SIZE_WIDTH = 110;
 
-    private AdapterInterface $cache;
+    private CacheItemPoolInterface $cache;
 
     private EntityManagerInterface $em;
 
-    public function __construct(AdapterInterface $cache, EntityManagerInterface $em)
+    public function __construct(CacheItemPoolInterface $cache, EntityManagerInterface $em)
     {
         $this->cache = $cache;
         $this->em = $em;
     }
 
-    /**
-     * @Route("/competences", name="cv-my_skills")
-     */
+    #[Route(path: '/competences', name: 'cv-my_skills')]
     public function my_skills(): Response
     {
         return $this->render('index/cv/my_skills.twig', [
@@ -47,33 +45,25 @@ class CuriculumVitaeController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/competences/logo/{slug}", name="cv-technology_icon")
-     */
+    #[Route(path: '/competences/logo/{slug}', name: 'cv-technology_icon')]
     public function technology_icon(Technologie $tech): Response
     {
         return AssetsResponse::ReturnImgAdapterCache($this->cache, $tech->getImage(), $tech->getSlug(), $tech->getImageExtension(), self::TECHNOLOGIES_SIZE_WIDTH);
     }
 
-    /**
-     * @Route("/cv", name="cv-my_career")
-     */
+    #[Route(path: '/cv', name: 'cv-my_career')]
     public function my_career(): Response
     {
         return $this->render('index/cv/my_cv.twig', []);
     }
 
-    /**
-     * @Route("/passion", name="cv-my_passions")
-     */
+    #[Route(path: '/passion', name: 'cv-my_passions')]
     public function my_passion(): Response
     {
         return $this->render('index/cv/passion.html.twig', []);
     }
 
-    /**
-     * @Route("/realisations", name="cv-making")
-     */
+    #[Route(path: '/realisations', name: 'cv-making')]
     public function making(RealisationRepository $realisationRepository): Response
     {
         if ($this->isGranted('ROLE_PREVIEW_MAKING')) {
@@ -89,17 +79,15 @@ class CuriculumVitaeController extends AbstractController
                 $this->cache->save($i);
             }
         }
-
         /** @var Realisation[] $makings */
         $makings = $i->get();
-
         return $this->render('index/cv/making.twig', ['realisations' => $makings]);
     }
 
     /**
-     * @Route("/realisation/img/{rea}", name="cv-making-img")
      * @ParamConverter("_r", options={"mapping": {"rea": "id"}})
      */
+    #[Route(path: '/realisation/img/{rea}', name: 'cv-making-img')]
     public function making_image(RealisationImage $_r): Response
     {
         if (!$this->isGranted('ROLE_PREVIEW_MAKING')) {
@@ -109,13 +97,10 @@ class CuriculumVitaeController extends AbstractController
                 }
             }
         }
-
-        return AssetsResponse::ReturnImgAdapterCache($this->cache, $_r->getImage(), (string) $_r->getId(), null, 850);
+        return AssetsResponse::ReturnImgAdapterCache($this->cache, $_r->getImage(), (string)$_r->getId(), null, 850);
     }
 
-    /**
-     * @Route("/realisation/{slug}", name="cv-making-view")
-     */
+    #[Route(path: '/realisation/{slug}', name: 'cv-making-view')]
     public function making_view(string $slug, RealisationRepository $realisationRepository): Response
     {
         $i = $this->cache->getItem(StringHelper::strRemoveCacheKey(self::CACHE_KEY_REALISATION_VIEW . $slug));
@@ -123,19 +108,16 @@ class CuriculumVitaeController extends AbstractController
             $i->set($realisationRepository->findBySlug($slug));
             $this->cache->save($i);
         }
-
         /** @var Realisation|null $v */
         $v = $i->get();
         if (null === $v || (!$v->isPublic() && !$this->isGranted('ROLE_PREVIEW_MAKING'))) {
             throw $this->createNotFoundException();
         }
-
         $alreadyDisplay = array_reduce($v->getTechnologies()->getValues(), function (array $acc, Technologie $t) {
             $acc[] = $t->getId();
 
             return $acc;
         }, []);
-
         $i2 = $this->cache->getItem(StringHelper::strRemoveCacheKey(self::CACHE_KEY_REALISATION_VIEW_LINKE_TECH . $slug));
         if (!$i2->isHit() || !$i->isHit()) {
             $i2->set(array_reduce($v->getTechnologies()->getValues(), function (array $acc, Technologie $t) use ($alreadyDisplay) {
@@ -149,19 +131,15 @@ class CuriculumVitaeController extends AbstractController
             }, []));
             $this->cache->save($i);
         }
-
         /** @var Technologie $linkTechno */
         $linkTechno = $i2->get();
-
         $i = $this->cache->getItem(StringHelper::strRemoveCacheKey(self::CACHE_KEY_REALISATION_VIEW_GALLERY . $slug));
         if (!$i->isHit()) {
             $i->set($this->em->getRepository(RealisationImageGallerie::class)->findBy(['realisations' => $v->getId()]));
             $this->cache->save($i);
         }
-
         /** @var RealisationImageGallerie[] $gallery */
         $gallery = $i->get();
-
         return $this->render('index/cv/making_view.twig', [
             'realisation' => $v,
             'gallery' => $gallery,
@@ -169,19 +147,16 @@ class CuriculumVitaeController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/ui")
-     */
+    #[Route(path: '/ui')]
     public function ui(): Response
     {
         return $this->render('index/ui.html.twig');
     }
 
     /**
-     * @Route("/by")
-     *
      * @return Response
      */
+    #[Route(path: '/by')]
     public function by()
     {
         return new Response('Fait par Ludwig SILVAIN');
