@@ -2,37 +2,33 @@
 
 namespace App\Entity\Main\CV;
 
+use Doctrine\ORM\Mapping\Entity;
+use Doctrine\ORM\Mapping\Table;
+use Doctrine\ORM\Mapping\Id;
+use Doctrine\ORM\Mapping\GeneratedValue;
+use Doctrine\ORM\Mapping\Column;
+use Symfony\Component\Validator\Constraints\Length;
+use Symfony\Component\Validator\Constraints\NotBlank;
 use App\Repository\Main\CV\CompetenceNiveauRepository;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Validator\Constraints as Assert;
 
-/**
- * @ORM\Entity(repositoryClass=CompetenceNiveauRepository::class)
- * @ORM\Table(name="CV_CompetenceNiveau")
- */
+#[Entity(repositoryClass: CompetenceNiveauRepository::class)]
+#[Table(name: 'CV_CompetenceNiveau')]
 class CompetenceNiveau
 {
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
+    #[Id]
+    #[GeneratedValue]
+    #[Column(type: 'integer')]
     private int $id;
-
-    /**
-     * @ORM\Column(type="string", length=32, name="NomNiveau")
-     * @Assert\Length(max="32")
-     * @Assert\NotBlank
-     */
+    #[Column(name: 'NomNiveau', type: 'string', length: 32)]
+    #[Length(max: 32)]
+    #[NotBlank]
     private string $name;
-
-    /**
-     * @ORM\Column(type="string", length=10, name="ClassNiveau")
-     * @Assert\Length(max="10")
-     * @Assert\NotBlank
-     */
+    #[Column(name: 'ClassNiveau', type: 'string', length: 10)]
+    #[Length(max: 10)]
+    #[NotBlank]
     private string $class;
-
     public function getId(): ?int
     {
         if (!isset($this->id)) {
@@ -41,24 +37,20 @@ class CompetenceNiveau
 
         return $this->id;
     }
-
     public function getName(): ?string
     {
         return $this->name;
     }
-
     public function setName(string $name): self
     {
         $this->name = $name;
 
         return $this;
     }
-
     public function getClass(): string
     {
         return $this->class;
     }
-
     public function setClass(string $class): self
     {
         $this->class = $class;
