@@ -2,6 +2,10 @@
 
 namespace App\Entity\Main;
 
+use Doctrine\ORM\Mapping\Entity;
+use Doctrine\ORM\Mapping\Id;
+use Doctrine\ORM\Mapping\GeneratedValue;
+use Doctrine\ORM\Mapping\Column;
 use DateTime;
 use DateTimeInterface;
 use Doctrine\ORM\Mapping as ORM;
@@ -9,32 +13,23 @@ use geertw\IpAnonymizer\IpAnonymizer;
 
 /**
  * Class ContactLog.
- *
- * @ORM\Entity(repositoryClass="App\Repository\Main\ContactLogRepository")
  */
+#[Entity(repositoryClass: 'App\Repository\Main\ContactLogRepository')]
 class ContactLog
 {
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     * @ORM\Column(type="integer")
-     */
+    #[Id]
+    #[GeneratedValue(strategy: 'AUTO')]
+    #[Column(type: 'integer')]
     protected int $id;
-    /**
-     * @ORM\Column(type="string", length=32)
-     */
+    #[Column(type: 'string', length: 32)]
     protected string $ipAddress;
-    /**
-     * @ORM\Column(type="datetime")
-     */
+    #[Column(type: 'datetime')]
     protected DateTimeInterface $dateCreated;
-
     public function __construct(string $ipAddress)
     {
         $this->dateCreated = new DateTime();
         $this->ipAddress = (new IpAnonymizer())->anonymize($ipAddress);
     }
-
     public function getDateCreated(): DateTimeInterface
     {
         return $this->dateCreated;
