@@ -2,49 +2,39 @@
 
 namespace App\Entity\Main;
 
+use Doctrine\ORM\Mapping\Entity;
+use Doctrine\ORM\Mapping\HasLifecycleCallbacks;
+use Doctrine\ORM\Mapping\Id;
+use Doctrine\ORM\Mapping\GeneratedValue;
+use Doctrine\ORM\Mapping\Column;
+use Doctrine\ORM\Mapping\PrePersist;
 use App\Repository\Main\UtilisateurRepository;
 use DateTime;
 use DateTimeInterface;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 
-/**
- * @ORM\Entity(repositoryClass=UtilisateurRepository::class)
- * @ORM\HasLifecycleCallbacks
- */
-class Utilisateur implements UserInterface
+#[Entity(repositoryClass: UtilisateurRepository::class)]
+#[HasLifecycleCallbacks]
+class Utilisateur implements UserInterface, PasswordAuthenticatedUserInterface
 {
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer", name="CodeUtilisateur")
-     */
+    #[Id]
+    #[GeneratedValue]
+    #[Column(name: 'CodeUtilisateur', type: 'integer')]
     private int $id;
-
-    /**
-     * @ORM\Column(type="string", length=180, unique=true, name="Utilisateur")
-     */
+    #[Column(name: 'Utilisateur', type: 'string', length: 180, unique: true)]
     private string $username;
-
-    /**
-     * @ORM\Column(type="string", length=180, unique=true, name="EmailUtilisateur")
-     */
+    #[Column(name: 'EmailUtilisateur', type: 'string', length: 180, unique: true)]
     private string $email;
-
-    /**
-     * @ORM\Column(type="json")
-     */
+    #[Column(type: 'json')]
     private array $roles = [];
-
     /**
      * @var string The hashed password
-     * @ORM\Column(type="string")
      */
+    #[Column(type: 'string')]
     private string $password;
-
-    /**
-     * @ORM\Column(type="datetime", name="DateCreationUtilisateur")
-     */
+    #[Column(name: 'DateCreationUtilisateur', type: 'datetime')]
     private DateTimeInterface $date_creation;
 
     public function getId(): ?int
@@ -75,7 +65,7 @@ class Utilisateur implements UserInterface
      */
     public function getUsername(): string
     {
-        return (string) $this->username;
+        return $this->username;
     }
 
     public function setUsername(string $username): self
@@ -109,7 +99,7 @@ class Utilisateur implements UserInterface
      */
     public function getPassword(): string
     {
-        return (string) $this->password;
+        return $this->password;
     }
 
     public function setPassword(string $password): self
@@ -148,11 +138,14 @@ class Utilisateur implements UserInterface
         return $this;
     }
 
-    /**
-     * @ORM\PrePersist
-     */
+    #[PrePersist]
     public function setCreatedValue(): void
     {
         $this->date_creation = new DateTime();
+    }
+
+    public function getUserIdentifier(): string
+    {
+        return $this->getUsername();
     }
 }
