@@ -2,6 +2,7 @@
 
 namespace App\Utils;
 
+use InvalidArgumentException;
 class StringHelper
 {
     private const unwanted_array_accent = ['Š' => 'S', 'š' => 's', 'Ž' => 'Z', 'ž' => 'z', 'À' => 'A', 'Á' => 'A', 'Â' => 'A', 'Ã' => 'A', 'Ä' => 'A', 'Å' => 'A', 'Æ' => 'A', 'Ç' => 'C', 'È' => 'E', 'É' => 'E',
@@ -29,7 +30,7 @@ class StringHelper
     public static function GenreateRandomString(int $length): string
     {
         if ($length < 0) {
-            throw new \InvalidArgumentException('$length can\'t be negative.');
+            throw new InvalidArgumentException('$length can\'t be negative.');
         }
         if (0 == $length) {
             return '';
